fixed bug where users could delete entries from other users

+++ src/server/trpc/router/entries.ts
@@ -43,10 +43,12 @@ export const entriesRouter = router({
       });
     }),
   delete: protectedProcedure.input(z.string()).mutation(({ ctx, input }) => {
-    return ctx.prisma.entry.delete({
-      where: {
-        id: input,
-      },
-    });
+    if (ctx.session.user.id == input) {
+      return ctx.prisma.entry.delete({
+        where: {
+          id: input,
+        },
+      });
+    }
   }),
 });
